requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
pip-autoremove==0.9.1
Pybooru==4.2.2
python-dotenv==0.18.0
requests==2.25.1
urllib3==1.26.6
opencv-python~=4.5.2.54
tensorflow~=2.5.0
Pillow~=8.3.0